fix(Config): Updated Config to the Python Getter and Setter style
