+++ ops/check_contents.py
@@ -4,7 +4,7 @@ from typing import Dict
 
 from db.inserts import insert_course
 from fileops.download_course import download_course
-from model.config import Config
+from model.Config import Config
 from model.course import Course
 from moodlews.get_course_contents import get_course_contents
 from moodlews.get_courses import get_courses
@@ -38,23 +38,23 @@ def __add_new_courses(conn: Connection, state: Config, courses: Dict[int, str]) 
     # Add the new courses to the state and stored them in the DB
     for c_id in new_courses:
         # Create the course and populate ir
-        course = Course(c_id, courses[c_id], state.get_default_action() == "download")
+        course = Course(c_id, courses[c_id], state.default_action == "download")
         content = get_course_contents(state, c_id)
         course.read_json_contents(content)
         state.add_course(course)
 
         # DB
-        insert_course(conn, course, state.id())
+        insert_course(conn, course, state.id)
 
     # Download the courses or notify
-    if state.get_default_action() == 'notify':
+    if state.default_action == 'notify':
         # Notify
         logging.warning("New courses were found, please check which contents you want to download.")
-    elif state.get_default_action() == 'download':
+    elif state.default_action == 'download':
         # Download
         logging.info("Downloading")
-        for course_key in state.get_courses():
-            course = state.get_courses()[course_key]
+        for course_key in state.courses:
+            course = state.courses[course_key]
             # download_course(course, state)
     return new_courses
 
@@ -71,7 +71,7 @@ def __check_courses_differences(state: Config, courses_id_received: list):
     new_courses = []
 
     for c_id in courses_id_received:
-        if c_id not in state.get_courses():
+        if c_id not in state.courses:
             new_courses.append(c_id)
 
     return new_courses
